@@ -9,6 +9,10 @@
     "directory": "packages/webpack-config"
   },
   "scripts": {
+    "lint": "eslint .",
+    "watch": "tsc --watch",
+    "build": "tsc",
+    "prepare": "yarn build",
     "test": "npm run test:unit; npm run test:e2e",
     "test:unit": "jest --config jest/unit-test-config.json --rootDir .",
     "test:e2e:start": "EXPO_E2E_COMMAND='start' jest --config jest/e2e-test-config.json --rootDir .",
@@ -36,7 +40,7 @@
     "chalk": "^2.4.2",
     "clean-webpack-plugin": "^1.0.1",
     "compression-webpack-plugin": "^2.0.0",
-    "copy-webpack-plugin": "^5.0.0",
+    "copy-webpack-plugin": "5.0.0",
     "css-loader": "^2.1.1",
     "deep-diff": "^1.0.2",
     "file-loader": "^3.0.1",
@@ -51,11 +55,11 @@
     "postcss-safe-parser": "^4.0.1",
     "prettier": "^1.16.4",
     "progress-bar-webpack-plugin": "^1.12.1",
-    "react-dev-utils": "^7.0.3",
+    "react-dev-utils": "9.0.3",
     "style-loader": "^0.23.1",
     "terser-webpack-plugin": "^1.2.3",
     "url-loader": "^1.1.2",
-    "webpack": "4.24.0",
+    "webpack": "4.29.6",
     "webpack-bundle-analyzer": "^3.0.4",
     "webpack-deep-scope-plugin": "1.6.0",
     "webpack-manifest-plugin": "^2.0.4",
@@ -66,9 +70,23 @@
     "@babel/core": "^7.4.5",
     "@babel/preset-flow": "^7.0.0",
     "@expo/babel-preset-cli": "^0.2.1-alpha.0",
+    "@types/copy-webpack-plugin": "^5.0.0",
+    "@types/deep-diff": "^1.0.0",
+    "@types/html-webpack-plugin": "^3.2.0",
+    "@types/node": "^12.0.12",
+    "@types/optimize-css-assets-webpack-plugin": "^5.0.0",
+    "@types/react-dev-utils": "^9.0.1",
+    "@types/terser-webpack-plugin": "^1.2.1",
+    "@types/webpack": "^4.32.1",
+    "@types/webpack-bundle-analyzer": "^2.13.2",
+    "@types/webpack-dev-server": "^3.1.7",
+    "@types/webpack-manifest-plugin": "^2.0.0",
+    "@types/webpack-merge": "^4.1.5",
+    "@types/workbox-webpack-plugin": "^4.1.0",
     "jest-puppeteer": "^4.3.0",
     "puppeteer": "^1.19.0",
-    "serve": "^11.1.0"
+    "serve": "^11.1.0",
+    "typescript": "3.4.5"
   },
   "gitHead": "613642fe06827cc231405784b099cf71c29072df",
   "browserslist": {
