@@ -1,4 +1,7 @@
-module.exports = function normalizePaths(initial, transformString) {
+export default function normalizePaths(
+  initial: any,
+  transformString: (value: string) => string
+): any {
   if (initial == null) {
     return initial;
   } else if (typeof initial === 'string') {
@@ -6,7 +9,7 @@ module.exports = function normalizePaths(initial, transformString) {
   } else if (Array.isArray(initial)) {
     return initial.map(value => normalizePaths(value, transformString));
   } else if (typeof initial === 'object') {
-    let result = {};
+    let result: { [key: string]: any } = {};
     for (const prop of Object.keys(initial)) {
       result[prop] = normalizePaths(initial[prop], transformString);
     }
@@ -14,4 +17,4 @@ module.exports = function normalizePaths(initial, transformString) {
   } else {
     return initial;
   }
-};
+}
