@@ -1,7 +1,19 @@
 /**
  * This creates environment variables that won't be tree shaken.
  */
-module.exports = function createClientEnvironment(mode, publicUrl, nativeAppManifest) {
+import { ExpoConfig } from '@expo/config';
+import { Mode } from './types';
+
+export interface ClientEnv {
+  __DEV__: boolean;
+  'process.env': { [key: string]: string };
+}
+
+function createClientEnvironment(
+  mode: Mode,
+  publicPath: string,
+  nativeAppManifest: ExpoConfig
+): ClientEnv {
   const environment = mode || 'development';
   const __DEV__ = environment !== 'production';
 
@@ -23,21 +35,23 @@ module.exports = function createClientEnvironment(mode, publicUrl, nativeAppMani
 
         /**
          * Useful for resolving the correct path to static assets in `public`.
-         * For example, <img src={process.env.WEB_PUBLIC_URL + '/img/logo.png'} />.
+         * For example, <img src={process.env.PUBLIC_URL + '/img/logo.png'} />.
          * This should only be used as an escape hatch. Normally you would put
          * images into the root folder and `import` them in code to get their paths.
          */
-        WEB_PUBLIC_URL: JSON.stringify(publicUrl),
+        PUBLIC_URL: JSON.stringify(publicPath),
 
         /**
          * Surfaces the `app.json` (config) as an environment variable which is then parsed by
          * `expo-constants` https://docs.expo.io/versions/latest/sdk/constants/
          */
         APP_MANIFEST: JSON.stringify(nativeAppManifest),
-      }
+      } as { [key: string]: string }
     );
   return {
     'process.env': processEnv,
     __DEV__,
   };
-};
+}
+
+export default createClientEnvironment;
